# docs/frontend-reference/openapi.yaml
openapi: 3.0.3
info:
  title: Kebda Zaman API
  version: "0.1.0"
  description: >
    Hand-authored directly from the NestJS source (no @nestjs/swagger is installed in this
    backend, so this could not be tool-generated from decorators — see the final report in the
    conversation that produced this doc set). Every path, DTO, and status code below was verified
    against the actual controller/DTO/service source, not invented. Base server path is /api/v1.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponseDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        role: { type: string, enum: [CUSTOMER, ADMIN] }
        isGuest: { type: boolean }
        locale: { type: string }
        createdAt: { type: string, format: date-time }
...
    DeliveryAddressDto:
      type: object
      required: [title, street, building, city]
      properties:
        title: { type: string, maxLength: 100 }
        street: { type: string, maxLength: 200 }
        building: { type: string, maxLength: 100 }
        floor: { type: string, maxLength: 50 }
        apartment: { type: string, maxLength: 50 }
        city: { type: string, maxLength: 100 }
    CheckoutDto:
      type: object
      required: [deliveryMethod, paymentMethod]
      properties:
        deliveryMethod: { type: string, enum: [DELIVERY, PICKUP] }
        paymentMethod: { type: string, enum: [CASH, CARD, WALLET] }
        deliveryAddress: { $ref: '#/components/schemas/DeliveryAddressDto' }
        promoCode: { type: string, maxLength: 50, description: 'Mutually exclusive with redeemRewardId' }
        redeemRewardId: { type: string, maxLength: 50, description: 'One of the fixed loyalty reward ids; mutually exclusive with promoCode; redeemed atomically in the checkout transaction' }
        notes: { type: string, maxLength: 1000 }
    OrderItemResponseDto:
      type: object
      properties:
        id: { type: string }
        menuItem:
          type: object
          properties:
            nameAr: { type: string }
            nameEn: { type: string }
            imageUrl: { type: string, nullable: true }
        selectedVariant:
          type: object
          nullable: true
          properties:
            id: { type: string }
            nameAr: { type: string }
            nameEn: { type: string }
            priceSnapshot: { type: number }
        selectedAddons:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              nameAr: { type: string }
              nameEn: { type: string }
              priceSnapshot: { type: number }
        quantity: { type: integer }
        specialInstructions: { type: string, nullable: true }
        unitPrice: { type: number }
        totalPrice: { type: number }
    OrderResponseDto:
      type: object
      properties:
        id: { type: string }
        orderNumber: { type: string }
        userId: { type: string }
        user: { $ref: '#/components/schemas/UserResponseDto' }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItemResponseDto' }
        status: { type: string, enum: [pending, confirmed, preparing, outForDelivery, delivered, cancelled] }
        deliveryAddress: {}
        paymentMethod: { type: string, enum: [cash, card, wallet] }
        subtotal: { type: number }
        deliveryFee: { type: number }
        tax: { type: number }
        discount: { type: number }
        totalAmount: { type: number }
        createdAt: { type: string, format: date-time }
        estimatedDeliveryTime: { type: string, nullable: true, format: date-time }
        loyaltyRedemption:
          type: object
          nullable: true
          properties:
            rewardId: { type: string }
            rewardName: { type: string }
            pointsRedeemed: { type: integer }
    OrderStatusResponseDto:
      type: object
      properties:
        status: { type: string }
        statusHistory:
          type: array
          items:
            type: object
            properties:
              status: { type: string }
              note: { type: string, nullable: true }
              changedAt: { type: string, format: date-time }
        estimatedDeliveryTime: { type: string, nullable: true }
...
paths:
...
  /orders:
    post:
      summary: '[CUSTOMER] Alias of /checkout (same handler)'
      parameters: [{ name: Idempotency-Key, in: header, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CheckoutDto' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/OrderResponseDto' } } } } }
    get:
      summary: List the caller's own orders
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, confirmed, preparing, outForDelivery, delivered, cancelled] } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/OrderResponseDto' } } } } } }
  /orders/{id}:
    get:
      summary: Get one of the caller's own orders
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderResponseDto' } } } }
        '404': { description: Not found }
  /orders/{id}/status:
    get:
      summary: Get status + full history for one of the caller's own orders
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderStatusResponseDto' } } } }
        '404': { description: Not found }
...
  /admin/orders/{id}:
    get:
      summary: '[ADMIN] Get any order by id'
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderResponseDto' } } } }
        '404': { description: Not found }
...